requirements.txt
asgiref==3.11.1
Django==5.2.11
django-environ==0.12.0
django-nested-admin==4.1.6
mysql-connector-python
pillow==12.1.0
python-monkey-business==1.1.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3